Squashed commit of the following:
commit 90574f1de34ec23b5ef6489f8e8a9cdd0c443eb6
Author: ielbusmente <ieljbusmente@gmail.com>
Date:   Mon Aug 29 20:54:12 2022 +0800

    updated database
    add read patients
    setup patient report

commit 6722af40e0fe851bcf0caff5cfdaa55cace41747
Author: ielbusmente <ieljbusmente@gmail.com>
Date:   Mon Aug 29 16:34:17 2022 +0800

    added session id and barangay id
    fixed alert scripts
    modified redirect conditions
    added conditions when midwife add and edits midwife accounts
    added conditional rendering on view midwife page and view nurse page

commit a7f7dd551063b262a70c8b39a1f53f687a5dae73
Author: ielbusmente <ieljbusmente@gmail.com>
Date:   Mon Aug 29 14:29:36 2022 +0800

    bug fix

commit 60c8e3e83a3e5ed33d9f3b06973688a8de8ad73e
Author: ielbusmente <ieljbusmente@gmail.com>
Date:   Mon Aug 29 14:19:10 2022 +0800

    Squashed commit of the following:

    commit 5817cfba35a7531e0946e3df0e80004251f5bb9c
    Author: ielbusmente <ieljbusmente@gmail.com>
    Date:   Mon Aug 29 14:16:22 2022 +0800

        updated redirecting, navigation, and pages

    commit e69dddd88d6fdcf697827ef205f09de75fe0e4f2
    Author: ielbusmente <ieljbusmente@gmail.com>
    Date:   Sun Aug 28 19:42:22 2022 +0800

        merge from crud

    commit b8f417a4f16fa505e09ac887f948c417018ab142
    Author: ielbusmente <ieljbusmente@gmail.com>
    Date:   Sun Aug 28 19:39:31 2022 +0800

        conditional rendering setup

    commit a9fc3fab2ea2656d966839d841193c73d2f159e0
    Merge: 015df45 0da8ce7
    Author: Edward Daniel Busmente <edward.busmente@blackfortconsulting.com>
    Date:   Sun Aug 28 10:13:08 2022 +0800

        navigation merge

    commit 0da8ce71d1d3087ac332dfcc0de0f20e4a2d75c0
    Author: ielbusmente <ieljbusmente@gmail.com>
    Date:   Fri Aug 26 16:48:57 2022 +0800

        added all pages in navigation

    commit 29590365d4d3d306db674c7c8a74d329a8d3a542
    Author: ielbusmente <ieljbusmente@gmail.com>
    Date:   Wed Aug 24 22:27:08 2022 +0800

        added profile change password page

    commit 365c0e7d58103ebae263e11e2ab60c2adf90799c
    Author: ielbusmente <ieljbusmente@gmail.com>
    Date:   Wed Aug 24 22:16:45 2022 +0800

        name on header

+++ admin/profile/view-nurse.php
@@ -4,9 +4,9 @@
 
 session_start();
 
-if(!isset($_SESSION['usermail'])) 
-  header('location:../'); 
-
+@include '../php-templates/redirect/admin-page-setter.php';
+@include '../php-templates/redirect/nurse-only.php';
+ 
 // fetch nurses 
 $select = "SELECT * FROM users WHERE admin = 1";
 $result = mysqli_query($conn, $select);
@@ -34,7 +34,7 @@ else  {
 $conn->close(); 
 
 $page = 'view_nurse';
-// $additional_script = '<script defer src="../js/nurse-table.js"></script>';
+// $additional_script = '<script defer src="../js/nurse-table.js"></script>'; 
 include_once('../php-templates/admin-navigation-head.php');
 ?>
  
@@ -70,7 +70,9 @@ include_once('../php-templates/admin-navigation-head.php');
                   <td><?php echo $value['status']; ?></td>
                   <td>
                     <button class="edit"><a href="edit-nurse.php?id=<?php echo $value['id'] ?>">Edit</a></button>
-                    <button class="del"><a href="delete-nurse.php?id=<?php echo $value['id'] ?>">Delete</a></button> 
+                    <?php if ($_SESSION['id'] != $value['id']) { ?> 
+                      <button class="del"><a href="delete-nurse.php?id=<?php echo $value['id'] ?>">Delete</a></button> 
+                    <?php } ?> 
                   </td>
                 </tr>
               <?php 
@@ -79,7 +81,7 @@ include_once('../php-templates/admin-navigation-head.php');
             </tbody>
           </table>
         </div>
-      </div>
+      </div> 
     </div>
   </div>
 </div>
